@@ -10,9 +10,11 @@
         }
     ],
     "require": {
-        "php": ">=7.3",
-
+        "php": "^7.4 || ^8.0",
         "symfony/console": "^5.0"
+    },
+    "require-dev": {
+        "phpunit/phpunit": "7.5.17"
     },
     "autoload": {
         "psr-4": {
